Rename files in testing
