+++ tests/integration_tests/test_pytorch_distributed.py
@@ -9,9 +9,9 @@ import torch.distributed as dist
 
 import optuna
 from optuna.integration import TorchDistributedTrial
-from optuna.testing.pruner import DeterministicPruner
-from optuna.testing.storage import STORAGE_MODES
-from optuna.testing.storage import StorageSupplier
+from optuna.testing.pruners import DeterministicPruner
+from optuna.testing.storages import STORAGE_MODES
+from optuna.testing.storages import StorageSupplier
 
 
 @pytest.fixture(scope="session", autouse=True)
